requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
twilio==9.2.3
openai==1.54.3
requests==2.32.3
psutil==6.0.0
python-dotenv==1.0.1
jinja2==3.1.4
